+++ packages/saas-ui-assets/src/index.ts
@@ -1,0 +1,2 @@
+export { SaasUILogo, type SaasUILogoProps } from './logo'
+export { SaasUIIcon, type SaasUIIconProps } from './icon'