+++ packages/saas-ui-core/src/index.ts
@@ -114,6 +114,20 @@ export {
   useSaas,
 } from './provider'
 export {
+  Navbar,
+  NavbarBrand,
+  type NavbarBrandProps,
+  NavbarContent,
+  type NavbarContentProps,
+  NavbarItem,
+  type NavbarItemProps,
+  NavbarLink,
+  type NavbarProps,
+  NavbarProvider,
+  useNavbar,
+  useNavbarContext,
+} from './navbar'
+export {
   Nav,
   NavGroup,
   NavGroupContent,
@@ -209,3 +223,4 @@ export {
   ViewOffIcon,
 } from './icons'
 export { createIcon } from './icons/create-icon'
+export { type IconBadgeProps, IconBadge } from './icon-badge'