+++ packages/saas-ui-core/src/property/property.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
 import { render, testStories } from '@saas-ui/test-utils'
-import * as stories from '../stories/property.stories'
+import * as stories from './property.stories'
 
 const { Basic } = testStories<typeof stories>(stories)