+++ packages/saas-ui-forms/src/submit-button.tsx
@@ -29,9 +29,10 @@ export interface SubmitButtonProps extends ButtonProps {
 export const SubmitButton = forwardRef<SubmitButtonProps, 'button'>(
   (props, ref) => {
     const {
+      variant = 'primary',
       children = 'Submit',
-      disableIfUntouched,
-      disableIfInvalid,
+      disableIfUntouched = false,
+      disableIfInvalid = false,
       isDisabled: isDisabledProp,
       isLoading,
       ...rest
@@ -50,6 +51,7 @@ export const SubmitButton = forwardRef<SubmitButtonProps, 'button'>(
         {...rest}
         {...field}
         ref={ref}
+        variant={variant}
         type="submit"
         isLoading={formState.isSubmitting || isLoading}
         isDisabled={isDisabled}
@@ -59,11 +61,5 @@ export const SubmitButton = forwardRef<SubmitButtonProps, 'button'>(
     )
   }
 )
-
-SubmitButton.defaultProps = {
-  variant: 'primary',
-  disableIfUntouched: false,
-  disableIfInvalid: false,
-}
 
 SubmitButton.displayName = 'SubmitButton'