+++ packages/saas-ui-clerk/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @saas-ui/clerk
 
+## 3.0.18
+
+### Patch Changes
+
+- @saas-ui/auth@2.2.2
+
+## 3.0.17
+
+### Patch Changes
+
+- @saas-ui/auth@2.2.1
+
+## 3.0.16
+
+### Patch Changes
+
+- Updated dependencies [a3180b02]
+  - @saas-ui/auth@2.2.0
+
 ## 3.0.15
 
 ### Patch Changes